+++ pi_c_openmpExplicit.c
@@ -1,7 +1,7 @@
 /*
  *  A C program to calculate Pi using quadrature as an OpenMP annotated algorithm.
  *
- *  Copyright © 2008-9 Russel Winder
+ *  Copyright © 2008-10 Russel Winder
  */
 
 #include <stdio.h>
@@ -10,10 +10,10 @@
 
 void execute ( const int numberOfThreads ) {
   const long n = 1000000000l ;
-  const long double delta = 1.0 / n ;
+  const double delta = 1.0 / n ;
   const long long startTimeMicros = microsecondTime ( ) ;
   const long sliceSize =  n / numberOfThreads ;
-  long double sum = 0.0 ;
+  double sum = 0.0 ;
   int i ;
 #pragma omp parallel for private ( i ) reduction ( + : sum )
   for ( i = 0 ; i < numberOfThreads ; ++i ) {
@@ -21,15 +21,15 @@ void execute ( const int numberOfThreads ) {
     const long end = ( i + 1 ) * sliceSize ;
     long j ;
     for ( j = start ; j <= end ; ++j ) {
-      const long double x = ( j - 0.5 ) * delta ;
+      const double x = ( j - 0.5 ) * delta ;
       sum += 1.0 / ( 1.0 + x * x ) ;
     }
   }
-  const long double pi = 4.0 * sum * delta ;
-  const long double elapseTime = ( microsecondTime ( ) - startTimeMicros ) / 1e6 ;
-  printf ( "==== C OpenMP Explicit pi = %.25Lf\n" , pi ) ;
+  const double pi = 4.0 * sum * delta ;
+  const double elapseTime = ( microsecondTime ( ) - startTimeMicros ) / 1e6 ;
+  printf ( "==== C OpenMP Explicit pi = %.18lf\n" , pi ) ;
   printf ( "==== C OpenMP Explicit iteration count = %ld\n" ,  n ) ;
-  printf ( "==== C OpenMP Explicit elapse = %Lf\n" , elapseTime ) ;
+  printf ( "==== C OpenMP Explicit elapse = %lf\n" , elapseTime ) ;
   printf ( "==== C OpenMP Explicit threadCount = %d\n" , numberOfThreads ) ;
   printf ( "==== C OpenMP Explicit processor count = %d\n" , omp_get_num_procs ( ) ) ;
 }
