+++ pi_c_pthreadGlobalState.c
@@ -1,7 +1,7 @@
 /*
  *  A C program to calculate Pi using quadrature as a threads-based algorithm.
  *
- *  Copyright © 2009 Russel Winder
+ *  Copyright © 2009-10 Russel Winder
  */
 
 #include <stdio.h>
@@ -9,20 +9,20 @@
 #include "microsecondTime.h"
 
 #define n 1000000000l
-const long double delta = 1.0 / n ;
+const double delta = 1.0 / n ;
 
 long sliceSize  ;
 
-long double sum ;
+double sum ;
 pthread_mutex_t sumMutex ;
 
 void * partialSum ( void *const arg  ) {
   const long start = 1 + ( (int) arg ) * sliceSize ;
   const long end = ( ( (int) arg ) + 1 ) * sliceSize ;
-  long double localSum = 0.0 ;
+  double localSum = 0.0 ;
   long i ;
   for ( i = start ; i <= end ; ++i ) {
-    const long double x = ( i - 0.5 ) * delta ;
+    const double x = ( i - 0.5 ) * delta ;
     localSum += 1.0 / ( 1.0 + x * x ) ;
   }
   pthread_mutex_lock ( &sumMutex ) ;
@@ -46,11 +46,11 @@ void execute ( const int numberOfThreads ) {
   pthread_attr_destroy ( &attributes ) ;
   int status ;
   for ( i = 0 ; i < numberOfThreads ; ++i ) { pthread_join ( threads[i] , (void **) &status ) ; }
-  const long double pi = 4.0 * sum * delta ;
-  const long double elapseTime = ( microsecondTime ( ) - startTimeMicros ) / 1e6 ;
-  printf ( "==== C PThread global pi = %.25Lf\n" , pi ) ;
+  const double pi = 4.0 * sum * delta ;
+  const double elapseTime = ( microsecondTime ( ) - startTimeMicros ) / 1e6 ;
+  printf ( "==== C PThread global pi = %.18lf\n" , pi ) ;
   printf ( "==== C PThread global iteration count = %ld\n" ,  n ) ;
-  printf ( "==== C PThread global elapse = %Lf\n" , elapseTime ) ;
+  printf ( "==== C PThread global elapse = %lf\n" , elapseTime ) ;
   printf ( "==== C PThread global thread count = %d\n" , numberOfThreads ) ;
 }
